Added private/shared handouts, added main server page with links to map and handouts
The handout is shared by default, which is signaled with a star "*" prefix in
the scene tree and will appear in the served main page.

By editing the name in the tree and removing the star the handout will become
private. The star can be added and the handout will become shared again.

Handouts (and other assets) can be opened locally by activating the tree item
(doubleclick or return)

Files:

qtvtt.py
- Added locally opening assets with the preferred app (tokens, images, tracks,
  handouts)
- Added main page to the http server with links to the handouts and map
- Added serving files via http and removed serving random files, only shared
  handouts are allowed to be served.
- Added importing handouts (random archives) into the scene. Those handouts can
  be renamed and shared with the players via the http server's main page
- Added saving and loading handouts from the scene qvt file
- Expanded top level tree items by default (excep walls and doors since they can
  have tens or hundreths of elements)
- Renamed shared server and app thread img_bytes to g_img_bytes

README.md
- Added features, screenshot

Tests:
- Imported handouts, edited handout name, made shared/private
- Verified the server main page serves map and handouts properly, skipping
  private handouts
- Verified scene assets can be locally loaded with the preferred app