Minor fixes: Set cell anchor wrt image topleft, cursor position on markdown header token
Scenes with map images at non-zero offset (eg most of those imported from
Dungeon Scrawl) were displaying the grid offsetted wrt the map, since the anchor
(the absolute position where the grid starts drawing) wasn't taking the map
image position into account, only the cell offset plus 0,0.

Now when setting the cell diameter and offset, the anchor looks at the available
images and uses the most topleft coordinate (min of tops and min of lefts) to
offset the cell anchor.

When using markdown header token # ## ###, etc, the cursor position was being
left in the ruler line instead of the empty line. This seems to be some Qt issue
with the cursor position on empty lines, since this works fine if the line is
not empty.

The easy solution is to set the format before deleting the space token separator
and then delete the space char after the fact.

Files:

qtvtt.py
- Added list_getat utility function similar to dict.get, currently unused.
- Assorted comment updating
- parseFormat now deletes the markdown token first and the token space separator
  second, so the cursor stays above the ruler line even if it's an empty line
- setCellOffsetAndDiameter now looks at the existing images to set the cell
  anchor (start of grid drawing). Also moved inside populateGrid for
  encapsulation

Tests:
- Loaded qvt files from imported Dungeon Scrawl, the grid is now properly
  centered around the map instead of with 0,0 offset
- Edited texts using the markdown header token # ## ### etc, verified the cursor
  is left above the ruler in both empty and non empty lines